Refactor move components to web-components if they contain web components
