+++ app/components/index.ts
@@ -1,7 +1,4 @@
 export * from "./accent-color-picker";
-export * from "./dimension";
 export * from "./direction-switch";
-export * from "./logo";
-export * from "./project-file-transfer";
 export * from "./theme-selector";
 export * from "./transparency-toggle";
