+++ app/icons/editor-logo.tsx
