+++ app/site-utilities/index.ts
@@ -1,3 +1,2 @@
 export * from "./definitions";
-export * from "./icons";
 export * from "./schemas";
