+++ app/web-components/dimension/dimension.props.ts
