+++ app/web-components/logo/logo.props.ts
